requirements.txt
flask
flask-cors
google-generativeai
langdetect
sqlite3-binary
requests
gunicorn
